+++ History.md
@@ -2,6 +2,7 @@ unreleased
 ==========
 
   * Add `deprecate.function(fn, message)`
+  * Add `process.on('deprecation', fn)` emitter
   * Automatically generate message when omitted from `deprecate()`
 
 0.0.1 / 2014-06-15